WIP: Refactored parser data structure as it got too big
