+++ src/note/note-question-parser.ts
@@ -7,8 +7,9 @@ import { CardFrontBack, CardFrontBackUtil } from "src/card/questions/question-ty
 import { DataStore } from "src/data-stores/base/data-store";
 import { TopicPath, TopicPathList } from "src/deck/topic-path";
 import { frontmatterTagPseudoLineNum, ISRFile } from "src/file";
-import { QuestionParser } from "src/parser/parser";
-import { ParsedQuestionInfo, ParserOptions } from "src/parser/parser-data-structure";
+import ParsedCardInfo from "src/utils/parsers/data-structures/parser/parsed-card-info";
+import ParserOptions from "src/utils/parsers/data-structures/parser/parser-options";
+import { CardParser } from "src/utils/parsers/card-parser";
 import { SettingsUtil, SRSettings } from "src/settings";
 import {
     splitNoteIntoFrontmatterAndContent,
@@ -110,7 +111,7 @@ export class NoteQuestionParser {
         this.tagCacheList = tagCacheList;
 
         const result: Question[] = [];
-        const parsedQuestionInfoList: ParsedQuestionInfo[] = this.parseQuestions();
+        const parsedQuestionInfoList: ParsedCardInfo[] = this.parseQuestions();
         for (const parsedQuestionInfo of parsedQuestionInfoList) {
             const question: Question = this.createQuestionObject(parsedQuestionInfo, textDirection);
 
@@ -150,7 +151,7 @@ export class NoteQuestionParser {
         return result;
     }
 
-    private parseQuestions(): ParsedQuestionInfo[] {
+    private parseQuestions(): ParsedCardInfo[] {
         const settings = this.settings;
         const parserOptions: ParserOptions = {
             singleLineCardSeparator: settings.singleLineCardSeparator,
@@ -163,11 +164,11 @@ export class NoteQuestionParser {
         };
 
         // We pass contentText which has the frontmatter blanked out; see extractFrontmatter for reasoning
-        return QuestionParser.parse(this.noteFile.path, this.contentText, parserOptions);
+        return CardParser.parse(this.noteFile.path, this.contentText, parserOptions);
     }
 
     private createQuestionObject(
-        parsedQuestionInfo: ParsedQuestionInfo,
+        parsedQuestionInfo: ParsedCardInfo,
         textDirection: TextDirection,
     ): Question {
         const questionContext: string[] = this.noteFile.getQuestionContext(
