+++ src/ui/obsidian-ui-components/content-container/settings-page/main-page.tsx
@@ -2,7 +2,7 @@ import { ButtonComponent, setIcon, Setting, SettingGroup } from "obsidian";
 
 import { t } from "src/lang/helpers";
 import SRPlugin from "src/main";
-import { QuestionParser } from "src/parser/parser";
+import { CardParser } from "src/utils/parsers/card-parser";
 import { SettingsPage } from "src/ui/obsidian-ui-components/content-container/settings-page/settings-page";
 import {
     getPageIcon,
@@ -26,7 +26,7 @@ export class MainPage extends SettingsPage {
         openPage: (pageType: SettingsPageType) => void,
         scrollListener: (scrollPosition: number) => void,
     ) {
-        super(pageContainerEl, plugin, pageType, () => {}, display, openPage, scrollListener);
+        super(pageContainerEl, plugin, pageType, () => { }, display, openPage, scrollListener);
 
         this.containerEl.addClass("sr-main-page");
 
@@ -170,7 +170,7 @@ export class MainPage extends SettingsPage {
                         .setValue(this.plugin.data.settings.showParserDebugMessages)
                         .onChange(async (value) => {
                             this.plugin.data.settings.showParserDebugMessages = value;
-                            QuestionParser.setDebugParser(
+                            CardParser.setDebugParser(
                                 this.plugin.data.settings.showParserDebugMessages,
                             );
                             await this.plugin.savePluginData();
