+++ src/utils/parsers/data-structures/cards/card-fragments/note-with-card-fragments.ts
@@ -1,10 +1,4 @@
-export type CardFragmentType =
-    | "ROUGE_SR_COMMENT"
-    | "ROUGE_MULTILINE_SEPARATOR"
-    | "MALFORMED_INLINE_CARD"
-    | "MALFORMED_MULTILINE_CARD_WITH_END_MARKER"
-    | "MALFORMED_MULTILINE_CARD"
-    | "ROUGE_MULTILINE_END_MARKER";
+import { CardFragment, CardFragmentType } from "src/utils/parser/data-structures/card-fragments/card-fragment";
 
 /**
  * A collection of notes with card fragments
@@ -84,14 +78,4 @@ export class NoteWithCardFragments {
     addFragment(cardFragment: CardFragment) {
         this.cardFragments.push(cardFragment);
     }
-}
-
-/**
- * A card fragment
- */
-export interface CardFragment {
-    text: string;
-    type: CardFragmentType;
-    startLineNum: number;
-    endLineNum: number;
 }
