+++ src/utils/parsers/data-structures/lines/html-comment.ts
@@ -1,0 +1,8 @@
+interface HTMLCommentSearchResultElement {
+    lineNumber: number;
+    startIndex: number;
+    endIndex: number;
+    text: string;
+}
+
+export default HTMLCommentSearchResultElement;
