+++ src/utils/parsers/data-structures/lines/line-data.ts
@@ -1,0 +1,65 @@
+
+import { CardType } from "src/card/questions/question";
+
+/**
+ * The line data class
+ *
+ * This class is responsible for storing the lines of the current note, the current line number,
+ * the current line, the current line trimmed without whitespace, the current line trimmed without whitespace at end,
+ * the inline separators, and the multiline separators.
+ *
+ * @class LineData
+ * @extends {LineData}
+ */
+export default class LineData {
+    lines: string[]; // The lines of the current note
+    currentLineNum: number; // The current line number
+    currentLine: string; // The current line
+    currentLineEndTrimmed: string; // The current line trimmed without whitespace at end
+    currentLineTrimmed: string; // The current line trimmed without whitespace
+    inlineSeparators: Array<{ separator: string; type: CardType }>; // The inline separators
+    multilineSeparators: Array<{ separator: string; type: CardType }>; // The multiline separators
+
+    /**
+     * Creates a new instance of LineData
+     *
+     * @param noteText - The note text
+     * @param inlineSeparators - The inline separators
+     * @param multilineSeparators - The multiline separators
+     */
+    constructor(
+        noteText: string,
+        inlineSeparators: Array<{ separator: string; type: CardType }>,
+        multilineSeparators: Array<{ separator: string; type: CardType }>,
+    ) {
+        this.reset(noteText, inlineSeparators, multilineSeparators);
+    }
+
+    /**
+     * Resets the line data
+     *
+     * @param noteText - The note text
+     * @param inlineSeparators - The inline separators
+     * @param multilineSeparators - The multiline separators
+     */
+    reset(
+        noteText: string,
+        inlineSeparators: Array<{ separator: string; type: CardType }>,
+        multilineSeparators: Array<{ separator: string; type: CardType }>,
+    ): void {
+        this.lines = noteText.replaceAll("\r\n", "\n").split("\n");
+        this.currentLineNum = -1;
+        this.currentLine = "";
+        this.currentLineEndTrimmed = "";
+        this.currentLineTrimmed = "";
+        this.inlineSeparators = inlineSeparators;
+        this.multilineSeparators = multilineSeparators;
+    }
+
+    setCurrentLine(lineNum: number): void {
+        this.currentLine = this.lines[lineNum];
+        this.currentLineNum = lineNum;
+        this.currentLineEndTrimmed = this.currentLine.trimEnd();
+        this.currentLineTrimmed = this.currentLineEndTrimmed.trimStart();
+    }
+}
