+++ src/utils/parsers/data-structures/parser/parser-options.ts
@@ -1,0 +1,11 @@
+interface ParserOptions {
+    singleLineCardSeparator: string;
+    singleLineReversedCardSeparator: string;
+    multilineCardSeparator: string;
+    multilineReversedCardSeparator: string;
+    multilineCardEndMarker: string | null;
+    clozePatterns: string[];
+    useAtomicClozes: boolean;
+}
+
+export default ParserOptions;
