+++ src/utils/parsers/data-structures/parser/parser-states.ts
@@ -1,0 +1,13 @@
+// All the states that the parser can be in, when parsing a note for cards
+export type ParserStates =
+    | "READY_TO_PARSE"
+    | "PARSE_LINE"
+    | "EMPTY_LINE"
+    | "TEXT"
+    | "HTML_COMMENT_START_OR_END"
+    | "HTML_COMMENT_MIDDLE"
+    | "SR_HTML_COMMENT"
+    | "CLOZE"
+    | "INLINE_CARD"
+    | "MULTILINE_SEPARATOR"
+    | "MULTILINE_END_MARKER";
