+++ tests/unit/parser/01_simple_tests/parser-debugger.test.ts
@@ -1,4 +1,4 @@
-import { QuestionParser } from "src/parser/parser";
+import { CardParser } from "src/utils/parsers/card-parser";
 
 import { parserOptions } from "../../helpers/unit-test-parser-helper";
 
@@ -6,9 +6,9 @@ describe("Parser debug messages", () => {
     test("Messages disabled", () => {
         // replace console error log with an empty mock function
         const logSpy = jest.spyOn(global.console, "log").mockImplementation(() => { });
-        QuestionParser.setDebugParser(false);
+        CardParser.setDebugParser(false);
 
-        QuestionParser.parse("", "", parserOptions);
+        CardParser.parse("", "", parserOptions);
         expect(logSpy).toHaveBeenCalledTimes(0);
 
         // restore original console error log
@@ -18,9 +18,9 @@ describe("Parser debug messages", () => {
     test("Messages enabled", () => {
         // replace console error log with an empty mock function
         const logSpy = jest.spyOn(global.console, "log").mockImplementation(() => { });
-        QuestionParser.setDebugParser(true);
+        CardParser.setDebugParser(true);
 
-        QuestionParser.parse("", "", parserOptions);
+        CardParser.parse("", "", parserOptions);
         expect(logSpy).toHaveBeenCalled();
 
         // restore original console error log
